refactor: separating batch collector to as isolated module
