+++ server/features/expenses/indexing.ts
@@ -1,4 +1,5 @@
-import { caseWhen, excluded, max, type AppDatabase, type BatchCollector } from '#server/lib/db';
+import { caseWhen, excluded, max, type AppDatabase } from '#server/lib/db';
+import type BatchCollector from '#server/lib/BatchCollector';
 import { blacklistSearchableText } from '#server/lib/expenseHelper';
 import { getMultiUserTextsHashes, getTextHash, getTrigrams, splitArray } from '#server/lib/utils';
 import { and, eq, desc, lt, inArray, count, SQL, isNotNull, min, gte } from 'drizzle-orm';
