+++ server/lib/sessions.ts
@@ -2,7 +2,8 @@ import type { Context } from './itty';
 import type { ProtectedContext } from './trpc';
 import { and, eq, gt, sql } from 'drizzle-orm';
 import { UAParser } from 'ua-parser-js';
-import { BatchCollector, maybeBatch, type AppDatabase } from './db';
+import type { AppDatabase } from './db';
+import BatchCollector, { maybeBatch } from './BatchCollector';
 import type { CookieHeaders } from './CookieHeaders';
 import { differenceInMinutes, addMinutes, differenceInDays, addDays, differenceInHours } from 'date-fns';
 import { loginAttemptsTable, sessionsTable, userDevicesTable, usersTable } from '../../db/schema';
